requirements.txt
baidu-aip==2.2.18.0
selenium==3.141.0
playsound==1.3.0
